+++ matrix-sdk/src/androidTest/java/org/matrix/androidsdk/lazyloading/RoomMembersTest.java
@@ -22,6 +22,7 @@ import org.junit.BeforeClass;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
+import org.matrix.androidsdk.MXSession;
 import org.matrix.androidsdk.RestClient;
 import org.matrix.androidsdk.common.CommonTestHelper;
 import org.matrix.androidsdk.common.TestApiCallback;
@@ -39,7 +40,7 @@ public class RoomMembersTest {
 
     @BeforeClass
     public static void init() {
-        RestClient.initUserAgent(InstrumentationRegistry.getContext());
+        MXSession.initUserAgent(InstrumentationRegistry.getContext(), null);
     }
 
     @Test