+++ CHANGES.rst
@@ -1,3 +1,20 @@
+Changes to Matrix Android SDK in 0.9.13 (2018-11-06)
+=======================================================
+
+Improvements:
+ - Add RTL support
+ - PermalinkUtils is now able to parse a permalink
+
+Bugfix:
+ - Fix crash when change visibility room (vector-im/riot-android#2679)
+ - Move `invite_room_state` to the UnsignedData object (vector-im/riot-android#2555)
+
+API Change:
+ - MXSession.initUserAgent() takes a second parameter for flavor description.
+
+Build:
+ - Treat some Lint warnings as errors
+
 Changes to Matrix Android SDK in 0.9.12 (2018-10-18)
 =======================================================